+++ src/normalizers/NormalizeContrastZeroMean.hpp
@@ -29,7 +29,7 @@ protected:
    virtual void ioParam_normalizeFromPostPerspective(enum ParamsIOFlag ioFlag);
 
    static void subtractOffsetAndNormalize(pvwdata_t * dataStartPatch, int weights_per_patch, float offset, float normalizer);
-   int accumulateSumAndSumSquared(pvwdata_t * dataPatchStart, int weights_in_patch, double * sum, double * sumsq);
+   int accumulateSumAndSumSquared(pvwdata_t * dataPatchStart, int weights_in_patch, float * sum, float * sumsq);
 
 private:
    int initialize_base();
